♻️refactor: [모의면접] 비디오 컴포넌트 분리
- Video.tsx 생성 후 컴포넌트 분리

KAN-170
